+++ packages/typespec-ts/static/static-helpers/multipartHelpers.ts
@@ -1,9 +1,11 @@
+import { NodeReadableStream } from "./platform-types.js";
+
 /**
  * Valid values for the contents of a binary file.
  */
 export type FileContents =
   | string
-  | NodeJS.ReadableStream
+  | NodeReadableStream
   | ReadableStream<Uint8Array>
   | Uint8Array
   | Blob;